streamlit/requirements.txt
streamlit
requests
psycopg2-binary  # PostgreSQL connection
